Add Administrator Panel login and file extrcation support. Refactor CNEXT functions.
